pyproject.toml
[project]
name = "sanpi-check"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "rmc8", email = "k@rmc-8.com" }
]
dependencies = [
    "langchain>=0.3.16",
    "langchain-ollama>=0.2.2",
    "python-dotenv>=1.0.1",
    "fire>=0.7.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sanpi_check"]

[tool.rye.scripts]
start = "python src/main.py"
